requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
python-dateutil
supabase
